requirements.txt
Flask==1.1.2
gunicorn==20.0.4
Flask-pymysql==0.2.3
Flask-QRcode==3.0.0
Jinja2==2.11.1
flask_recaptcha==0.4.2
Flask-Mail==0.9.1
Pillow==7.2.0
boto3==1.14.16
